rename shopclasses to shop, logger to uppercase in main
