+++ entities/people/Customer.java
@@ -1,6 +1,6 @@
 package people;
 
-import shopclasses.Receipt;
+import shop.Receipt;
 
 /*
  * Customer class represents customers of the shop and their shopping receipt.
