+++ entities/shop/Item.java
@@ -1,4 +1,4 @@
-package shopclasses;
+package shop;
 
 /*
  * Item class represents all the goods that are in the store.
