+++ entities/shop/MarketAsile.java
@@ -1,4 +1,4 @@
-package shopclasses;
+package shop;
 
 /*
  * Market class is a type of Asile that represents the open market area to display all the fresh produce.
