+++ entities/shop/Receipt.java
@@ -1,4 +1,4 @@
-package shopclasses;
+package shop;
 
 import java.util.ArrayList;
 import java.util.Date;
