+++ entities/shop/Supplier.java
@@ -1,4 +1,4 @@
-package shopclasses;
+package shop;
 
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
